+++ src/res/logo.ico
