+++ src/res/win_res.rc
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 #include <windows.h>
-#include "chsrc_version.h"
+#include "version.h"
 
 // 图标资源
 IDI_ICON1 ICON DISCARDABLE "logo.ico"
